gui2/game_version: Move report text generator to game_config ns for general use
